+++ perf/tests/test_timeit.py
@@ -44,7 +44,7 @@ class TestTimeit(unittest.TestCase):
                          r'Metadata:\n'
                          r'(- .*\n)+'
                          r'\n'
-                         r'Median \+- std dev: (?P<avg>[0-9.]+) ms \+- (?P<stdev>[0-9.]+) ms '
+                         r'Median \+- std dev: (?P<median>[0-9.]+) ms \+- (?P<stdev>[0-9.]+) ms '
                              r'\(2 samples; 1 warmup\)\n'
                          r'$',
                          stdout)
@@ -55,8 +55,8 @@ class TestTimeit(unittest.TestCase):
             self.assertTrue(MIN_SAMPLE <= value <= MAX_SAMPLE,
                             repr(value))
 
-        mean = float(match.group('avg'))
-        self.assertTrue(MIN_MEAN <= mean <= MAX_MEAN, mean)
+        median = float(match.group('median'))
+        self.assertTrue(MIN_MEAN <= median <= MAX_MEAN, median)
         stdev = float(match.group('stdev'))
         self.assertLessEqual(stdev, MAX_STDEV)
 
@@ -75,15 +75,15 @@ class TestTimeit(unittest.TestCase):
         self.assertEqual(proc.returncode, 0)
 
         match = re.match(r'^\.\.\n'
-                         r'Median \+- std dev: (?P<avg>[0-9.]+) ms'
+                         r'Median \+- std dev: (?P<median>[0-9.]+) ms'
                              r' \+- (?P<stdev>[0-9.]+) ms'
                          r'$',
                          stdout.rstrip())
         self.assertIsNotNone(match, repr(stdout))
 
         # Tolerate large differences on busy systems
-        mean = float(match.group('avg'))
-        self.assertTrue(MIN_MEAN <= mean <= MAX_MEAN, mean)
+        median = float(match.group('median'))
+        self.assertTrue(MIN_MEAN <= median <= MAX_MEAN, median)
 
         stdev = float(match.group('stdev'))
         self.assertLessEqual(stdev, MAX_STDEV)